refactor: modularize forge-app router into domain-specific modules
Split the monolithic 1680-line router.rs into modular route files:
- routes/state.rs: ForgeAppState and related implementations
- routes/tasks.rs: Task creation, listing, and WebSocket streaming
- routes/attempts.rs: Task attempt handlers with forge overrides
- routes/config.rs: Global forge configuration endpoints
- routes/projects.rs: Project-specific settings and profiles
- routes/omni.rs: Omni service status and notifications
- routes/agents.rs: Forge agents and neuron management
- routes/frontend.rs: Static file serving and documentation

The main router.rs is now a thin ~145-line composition layer that
imports and composes these modules. Also added build.rs to create
dummy frontend/dist directory for compilation.
